parameters

+++ src/training/rob_test_pipeline.py
@@ -258,7 +258,7 @@ def start_student_instruction_generation():
     '''Creates X amount of new instructions by a student for robert'''
     params = get_parameters()
     my_student = robert(finetuned_path=build_finetuned_path("student_24k_para"),
-                        is_student=True, context_amount=4, dtype="bfloat16")
+                        is_student=True, context_amount=4)
     for i in range(100):
         context = random.sample(params, random.randint(1, 3))
         my_student.set_context(context)
